jz_efficientnet_b7/requirements.txt
torch==1.11.0+cu113
pandas==1.3.4
torchvision==0.12.0+cu113
pycocotools==2.0.6
numpy==1.22.4
pillow==8.4.0
tqdm==4.62.3